+++ randovania/gui/dialog/trick_usage_popup.py
@@ -82,6 +82,8 @@ class TrickUsagePopup(QtWidgets.QDialog, Ui_TrickUsagePopup):
         trick_level = preset.configuration.trick_level
         if trick_level.minimal_logic:
             trick_usage_description = "Minimal Logic"
+        elif not trick_level.specific_levels:
+            trick_usage_description = "None"
         else:
             trick_usage_description = ", ".join(
                 sorted(
@@ -106,6 +108,9 @@ class TrickUsagePopup(QtWidgets.QDialog, Ui_TrickUsagePopup):
         self.button_box.rejected.connect(self.button_box_close)
 
         if trick_level.minimal_logic:
+            self.area_list_label.setText(
+                "This preset uses minimal logic, meaning all possible tricks may potentially be required."
+            )
             return
 
         # Update
@@ -122,6 +127,8 @@ class TrickUsagePopup(QtWidgets.QDialog, Ui_TrickUsagePopup):
                 used_tricks = _check_used_tricks(area, trick_resources, context)
                 if used_tricks:
                     lines.append(data_editor_href(region, area) + f"<br />{'<br />'.join(used_tricks)}</p>")
+        if not lines:
+            lines.append("This is a trickless preset and thus, no tricks were listed.")
 
         self.area_list_label.setText("".join(lines))