+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [10.1.0] - 2025-09-0?
 
-
+- Fixed: "View Expected Trick Usage" now works correctly with trickless and minimal logic presets.
 - Fixed: Hidden Pickups are not shown anymore in the "Pickup Hint Features" tab of a game.
 
 ### AM2R